Movie class to new location
