--- HoneyBadgers.Logic/Searcher.cs
@@ -1,38 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-
-
-namespace HoneyBadgers.Logic
-{
-    public static class Searcher
-    {
-        public static Dictionary<Movie,int> FindByName(Data db, string searchInput)
-        {
-            //TODO: Jak się pozbyć Dictionary?! :o 
-            var inputParts = searchInput.Split(" "); 
-            var results = new Dictionary<Movie, int>();
-            foreach (var itemDb in db.Movies)
-            {
-                var movieTitle = itemDb.Title.ToLower();
-                var precision = 0;
-                foreach (var partInput in inputParts)
-                {
-                    if (movieTitle.Contains(partInput.ToLower()))
-                    {
-                        precision += 1;
-                    }
-                }
-                
-                if (precision > 0)
-                {
-                    
-                    results.Add(itemDb,precision);
-                }
-            }
-
-            results.OrderBy(x => x.Value).ToDictionary(x => x.Key, x => x.Value);
-            return results;
-        }
-    }
-}
